Add flush link to index page and change flush to POST (#9)
This changes flush from a GET to a POST. It changes the state of the
server, so it should have been a POST all along.

This adds a "[Flush]" link to the index page for convenience. It saves
you from having to curl the API directly or shutting down and restarting
Kent.

This adds a kent-testpost command for using sentry-sdk and capturing a
message. This makes it easier to manually test Kent.