Rework logging to be more helpful (#32)
This changes what gets logged when Kent gets a request. Previously, it
would log all the HTTP headers. That's not helpful when you're debugging
issues in an application. This changes it to log sentry-sdk related
headers as well as the exception type/value or message from the payload.

This also logs the url for the event if you want to look at it.