Fix summary and summary logging (#91)
This fixes how Kent handles csp reports so that the Event is more
correct and so we can rely on event.summary for logging a summary of the
event to the console.

This also adds a datestamp for csp reports. It doesn't come from the
payload--it's generated when handling the report.

This additionally fixes HTTP view code to use event.summary rather than
have redundant code to determine the summary.